test: move budget floor sweep tests into tests/presses for collection
tests/default_presses.py is not matched by pytest's default test_*.py
discovery, so the new regression tests added there never ran under
`pytest tests/` (make test). Move them to a self-contained
tests/presses/test_budget_floor_sweep.py next to the other press
tests (same placement as the #264 regression tests) and restore
tests/default_presses.py to its original state.

Refs #264
